# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hush Service API
  version: 1.0.0
  description: Single-owner encrypted vault. CLI devices are read-only. All binary
    fields use canonical unpadded base64url. See crypto.md for mandatory AAD,
    algorithm, size, and nonce rules. The Worker validates structure, not
    cryptographic correctness.
servers:
  - url: /api/v1
paths:
...
  /devices:
    get:
      operationId: listDevices
      summary: List device metadata and last ciphertext/key fetch time
      security: *a3
      parameters: []
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeviceMetadata"
                required:
                  - devices
                additionalProperties: false
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: enrollDevice
      summary: Enroll a client public key and wrapped VEK
      security: *a3
      parameters:
        - name: If-Match
          in: header
          required: true
          description: Quoted vault revision, e.g. "3". Initialization requires "0".
          schema:
            type: string
            pattern: ^"(0|[1-9][0-9]*)"$
        - name: Origin
          in: header
          required: true
          description: Must equal configured ADMIN_ORIGIN.
          schema: *a2
        - name: X-Hush-Request
          in: header
          required: true
          schema:
            const: "1"
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
  /devices/self/key:
    get:
      operationId: getOwnDeviceKey
      summary: Read wrapped VEK for the authenticated CLI device
      security:
        - AccessClientId: []
          AccessClientSecret: []
      parameters: []
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceKey"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{id}/key:
    get:
      operationId: getDeviceKeyAsOwner
      summary: Read an active device wrap as owner
      security: *a3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceKey"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /devices/{id}:
    patch:
      operationId: renameDevice
      summary: Rename a device
      security: *a3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: true
          description: Quoted vault revision, e.g. "3". Initialization requires "0".
          schema:
            type: string
            pattern: ^"(0|[1-9][0-9]*)"$
        - name: Origin
          in: header
          required: true
          description: Must equal configured ADMIN_ORIGIN.
          schema: *a2
        - name: X-Hush-Request
          in: header
          required: true
          schema:
            const: "1"
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Rename"
  /devices/{id}/revoke:
    post:
      operationId: revokeDevice
      summary: Soft-revoke a device and remove its wrap; no request body
      security: *a3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: true
          description: Quoted vault revision, e.g. "3". Initialization requires "0".
          schema:
            type: string
            pattern: ^"(0|[1-9][0-9]*)"$
        - name: Origin
          in: header
          required: true
          description: Must equal configured ADMIN_ORIGIN.
          schema: *a2
        - name: X-Hush-Request
          in: header
          required: true
          schema:
            const: "1"
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    DeviceInput:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        name: &a13
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[^\x00-\x1f\x7f]+$
        type: &a14
          type: string
          enum:
            - cli
            - browser
        public_key: &a15
          type: string
          minLength: 43
          maxLength: 43
          pattern: ^[A-Za-z0-9_-]+$
        access_client_id: &a16
          anyOf:
            - type: string
              minLength: 1
              maxLength: 256
              pattern: ^[A-Za-z0-9_-]+\.access$
            - type: "null"
        vek_version: &a17
          type: integer
          minimum: 1
          maximum: 9007199254740990
        wrapped_vek: &a18
          type: object
          properties:
            format:
              type: number
              const: 1
            algorithm:
              type: string
              const: HPKE-X25519-HKDF-SHA256-AES-256-GCM
            enc:
              type: string
              minLength: 43
              maxLength: 43
              pattern: ^[A-Za-z0-9_-]+$
            ciphertext:
              type: string
              minLength: 64
              maxLength: 64
              pattern: ^[A-Za-z0-9_-]+$
          required:
            - format
            - algorithm
            - enc
            - ciphertext
          additionalProperties: false
      required:
        - id
        - name
        - type
        - public_key
        - access_client_id
        - vek_version
        - wrapped_vek
      additionalProperties: false
...
    Rename:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[^\x00-\x1f\x7f]+$
      required:
        - name
      additionalProperties: false
...
    DeviceMetadata:
      type: object
      properties:
        id: *a25
        name: *a13
        type: *a14
        public_key: *a15
        access_client_id: *a16
        vek_version: *a17
        created_at: *a10
        updated_at: *a10
        status: *a26
        last_seen_at:
          anyOf:
            - *a10
            - type: "null"
      required:
        - id
        - name
        - type
        - public_key
        - access_client_id
        - vek_version
        - created_at
        - updated_at
        - status
        - last_seen_at
      additionalProperties: false
    Revision:
      type: object
      properties:
        revision: *a23
      required:
        - revision
      additionalProperties: false
...
    DeviceKey:
      type: object
      properties:
        device_id: *a2
        vault_id: *a2
        vek_version: *a23
        wrapped_vek: *a18
      required:
        - device_id
        - vault_id
        - vek_version
        - wrapped_vek
      additionalProperties: false
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *a2
            request_id: *a2
          required:
            - code
            - request_id
          additionalProperties: false
      required:
        - error
      additionalProperties: false
  securitySchemes:
    AccessSession:
      type: apiKey
      in: cookie
      name: CF_Authorization
      description: Cloudflare Access human session. Edge forwards the signed
        assertion; Worker verifies it and matches OWNER_EMAIL.
    AccessClientId:
      type: apiKey
      in: header
      name: CF-Access-Client-Id
    AccessClientSecret:
      type: apiKey
      in: header
      name: CF-Access-Client-Secret